rename directory to use all lowercase
