+++ src/App.js
@@ -7,7 +7,7 @@ import ResetForm from './containers/ResetPassword';
 import ConfirmPasswordForm from './containers/ConfirmPasswordForm';
 import SignUp from './containers/SignUp';
 import Dashboard from './containers/Dashboard';
-import PrivateRoute from './Routes/PrivateRoute';
+import PrivateRoute from './routes/PrivateRoute';
 
 const App = () => (
   <BrowserRouter>
